+++ libft/output/ft_puthex_fd.c
@@ -6,13 +6,13 @@
 /*   By: ldel-val <ldel-val@student.42madrid.c     |  |           *           */
 /*                                                 \  '.___.;       +         */
 /*   Created: 2024/10/28 17:00:39 by ldel-val       '._  _.'   .        .     */
-/*   Updated: 2024/10/28 17:48:29 by ldel-val          ``                     */
+/*   Updated: 2024/10/28 20:17:09 by ldel-val          ``                     */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-int	ft_puthex_fd(size_t n, int fd)
+int	ft_puthex_fd(long unsigned int n, int fd)
 {
 	int	printed_bytes;
