+++ libft/output/ft_putuphex_fd.c
@@ -6,13 +6,13 @@
 /*   By: ldel-val <ldel-val@student.42madrid.c     |  |           *           */
 /*                                                 \  '.___.;       +         */
 /*   Created: 2024/10/28 17:00:39 by ldel-val       '._  _.'   .        .     */
-/*   Updated: 2024/10/28 18:36:44 by ldel-val          ``                     */
+/*   Updated: 2024/10/28 20:21:31 by ldel-val          ``                     */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-int	ft_putuphex_fd(size_t n, int fd)
+int	ft_putuphex_fd(long unsigned int n, int fd)
 {
 	int	printed_bytes;
